+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   * updated Gradle to version 2.8
   * fixed a classpath problem in the `localizer` task
     ([#71](https://github.com/jenkinsci/gradle-jpi-plugin/pull/71))
+  * allow to specify the HTTP port for the `server` task with the `jenkins.httpPort` project or system property
+    ([JENKINS-31881](https://issues.jenkins-ci.org/browse/JENKINS-31881))
 
 ## 0.15.0 (2016-02-18)